FUNCIONA! Com IRQ! Porém algumas vezes trava com 4 min, outras vezes não trava em mais de 1h de execução. Uso da macro STD_ITER_DELAY em após cada leitura ou escrita de registrador. Diminuição do delay em main() para agilizar os testes.

+++ Libraries/nrf_lib/CONFIG.h
@@ -12,7 +12,7 @@
 #define CONFIG_H_
 
 //macro: standard iteractive delay of about 500us
-#define STD_ITER_DELAY for (int i=0;i<0xee2;i++);
+#define STD_ITER_DELAY for (uint16_t i=0;i<0xee2;i++);
 
 void TimingDelay_Decrement(void);
 void Delay_us(uint32_t time_us);
